requirements.txt
pvporcupine==1.9.2
sounddevice
numpy
aio_pika
vosk
scipy
pvrecorder
piper-tts